Updated pyproject and changed name
