+++ backup_requirements.txt
@@ -27,6 +27,7 @@ Django==4.2.16
 executing==2.1.0
 fastjsonschema==2.20.0
 fonttools==4.54.1
+-e git+ssh://git@github.com/SnowHana/toelo.git@5f74ca75f5fd36f39bd5e78f38f0e0eedd2cf151#egg=footy&subdirectory=footy
 fqdn==1.5.1
 gitdb==4.0.12
 GitPython==3.1.44
