+++ pyproject.toml
@@ -1,0 +1,150 @@
+[tool.poetry]
+name = "toelo"
+version = "0.1.0"
+description = "Football data anaylsis"
+authors = ["Nautilus-Ubuntu <wujindaniel1011@gmail.com>"]
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = "^3.12"
+# toelo = { git = "https://github.com/SnowHana/toelo.git", rev = "5f74ca75f5fd36f39bd5e78f38f0e0eedd2cf151" }
+altair = "5.5.0"
+anyio = "4.6.2.post1"
+appnope = "0.1.4"
+argon2-cffi = "23.1.0"
+argon2-cffi-bindings = "21.2.0"
+arrow = "1.3.0"
+asgiref = "3.8.1"
+asttokens = "2.4.1"
+async-lru = "2.0.4"
+attrs = "24.2.0"
+babel = "2.16.0"
+beautifulsoup4 = "4.12.3"
+bleach = "6.1.0"
+blinker = "1.9.0"
+cachetools = "5.5.2"
+certifi = "2024.8.30"
+cffi = "1.17.1"
+charset-normalizer = "3.4.0"
+click = "8.1.8"
+comm = "0.2.2"
+contourpy = "1.3.0"
+cycler = "0.12.1"
+debugpy = "1.8.7"
+decorator = "5.1.1"
+defusedxml = "0.7.1"
+django = "4.2.16"
+executing = "2.1.0"
+fastjsonschema = "2.20.0"
+fonttools = "4.54.1"
+fqdn = "1.5.1"
+gitdb = "4.0.12"
+gitpython = "3.1.44"
+greenlet = "3.1.1"
+h11 = "0.14.0"
+httpcore = "1.0.6"
+httpx = "0.27.2"
+idna = "3.10"
+ipykernel = "6.29.5"
+ipython = "8.28.0"
+ipywidgets = "8.1.5"
+isoduration = "20.11.0"
+jedi = "0.19.1"
+jinja2 = "3.1.4"
+json5 = "0.9.25"
+jsonpointer = "3.0.0"
+jsonschema = "4.23.0"
+jsonschema-specifications = "2024.10.1"
+jupyter = "1.1.1"
+jupyter-console = "6.6.3"
+jupyter-events = "0.10.0"
+jupyter-lsp = "2.2.5"
+jupyter-client = "8.6.3"
+jupyter-core = "5.7.2"
+jupyter-server = "2.14.2"
+jupyter-server-terminals = "0.5.3"
+jupyterlab = "4.2.5"
+jupyterlab-pygments = "0.3.0"
+jupyterlab-server = "2.27.3"
+jupyterlab-widgets = "3.0.13"
+kiwisolver = "1.4.7"
+markdown-it-py = "3.0.0"
+markupsafe = "3.0.2"
+matplotlib = "3.9.2"
+matplotlib-inline = "0.1.7"
+mdurl = "0.1.2"
+mistune = "3.0.2"
+narwhals = "1.29.0"
+nbclient = "0.10.0"
+nbconvert = "7.16.4"
+nbformat = "5.10.4"
+nest-asyncio = "1.6.0"
+notebook = "7.2.2"
+notebook-shim = "0.2.4"
+numpy = "2.1.2"
+overrides = "7.7.0"
+packaging = "24.1"
+pandas = "2.2.3"
+pandocfilters = "1.5.1"
+parso = "0.8.4"
+pexpect = "4.9.0"
+pillow = "11.0.0"
+platformdirs = "4.3.6"
+plotly = "5.24.1"
+prometheus-client = "0.21.0"
+prompt-toolkit = "3.0.48"
+protobuf = "5.29.3"
+psutil = "6.1.0"
+psycopg = {extras = ["binary"], version = "^3.2.6"}
+psycopg2 = "2.9.10"
+ptyprocess = "0.7.0"
+pure-eval = "0.2.3"
+pyarrow = "19.0.1"
+pycparser = "2.22"
+pydeck = "0.9.1"
+pygments = "2.18.0"
+pyparsing = "3.2.0"
+python-dateutil = "2.9.0.post0"
+python-json-logger = "2.0.7"
+pytz = "2024.2"
+pyyaml = "6.0.2"
+pyzmq = "26.2.0"
+referencing = "0.35.1"
+requests = "2.32.3"
+rfc3339-validator = "0.1.4"
+rfc3986-validator = "0.1.1"
+rich = "13.9.4"
+rpds-py = "0.20.0"
+scipy = "1.14.1"
+send2trash = "1.8.3"
+setuptools = "75.2.0"
+six = "1.16.0"
+smmap = "5.0.2"
+sniffio = "1.3.1"
+soupsieve = "2.6"
+sqlalchemy = "2.0.36"
+sqlparse = "0.5.1"
+stack-data = "0.6.3"
+streamlit = "1.43.0"
+tenacity = "9.0.0"
+terminado = "0.18.1"
+tinycss2 = "1.3.0"
+toml = "0.10.2"
+tornado = "6.4.1"
+traitlets = "5.14.3"
+types-python-dateutil = "2.9.0.20241003"
+typing-extensions = "4.12.2"
+tzdata = "2024.2"
+uri-template = "1.3.0"
+urllib3 = "2.2.3"
+wcwidth = "0.2.13"
+webcolors = "24.8.0"
+webencodings = "0.5.1"
+websocket-client = "1.8.0"
+wheel = "0.45.1"
+widgetsnbextension = "4.0.13"
+
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
