+++ src/footy.egg-info/PKG-INFO
