+++ src/footy.egg-info/SOURCES.txt
