+++ src/footy.egg-info/dependency_links.txt
