+++ src/footy.egg-info/entry_points.txt
