+++ src/footy.egg-info/requires.txt
