+++ src/footy.egg-info/top_level.txt
