+++ src/toelo/player_elo/club_analysis.py
