+++ src/toelo/player_elo/init_sql.py
@@ -1,11 +1,11 @@
 import os
 from pathlib import Path
 from typing import Dict
-from footy.player_elo.database_connection import DATA_DIR, DATABASE_CONFIG
+from toelo.player_elo.database_connection import DATA_DIR, DATABASE_CONFIG
 
 # Make sure your environment uses psycopg 3, e.g. 'psycopg==3.1.8'
 # and your SQLAlchemy URL is 'postgresql+psycopg://...' not 'psycopg2'
-from footy.player_elo.game_validator import validate_games
+from toelo.player_elo.game_validator import validate_games
 from sqlalchemy import (
     create_engine,
     Column,
