+++ src/toelo/player_elo/old/copy_init_player_elo.py
