+++ src/toelo/player_elo/old/game_analysis_old.py
