+++ src/toelo/player_elo/old/init_sql_old.py
